Automatically and silently updates from server if a newer version is on server.
Every time app is started, checks server directory listing for a newer
version number. If there is one, download the new update files and
silently install update so the app is updated the next time it is
started.
Still deletes DB even when Database.sdf is not included in update
package.